fix: code review — RETURNING id 제거, || → ??, SMALLINT → INTEGER
- ConsolidatorGC._gcSearchEvents: RETURNING id 제거 (rowCount만 사용)
- MemoryManager.toolFeedback: search_event_id || null → ?? null
- memory.js: _searchEventId || null → ?? null
- migration-013: latency_ms SMALLINT → INTEGER (32s 초과 대비)

+++ lib/memory/migration-013-search-events.sql
@@ -15,7 +15,7 @@ CREATE TABLE IF NOT EXISTS agent_memory.search_events (
     result_count   SMALLINT     NOT NULL DEFAULT 0,
     l1_is_fallback BOOLEAN      NOT NULL DEFAULT FALSE,
     used_rrf       BOOLEAN      NOT NULL DEFAULT FALSE,
-    latency_ms     SMALLINT,
+    latency_ms     INTEGER,
     query_type     TEXT         CHECK (query_type IN ('keywords', 'text', 'topic', 'mixed')),
     filter_keys    TEXT[],
     created_at     TIMESTAMPTZ  NOT NULL DEFAULT NOW()
